rewrite mongo to use new dataset loader interface
