+++ test/metabase/driver/mongo/test_data.clj
@@ -1,63 +1,21 @@
 (ns metabase.driver.mongo.test-data
   "Functionality related to creating / loading a test database for the Mongo driver."
-  (:require [clojure.tools.logging :as log]
-            [colorize.core :as color]
-            [medley.core :as m]
-            (monger [collection :as mc]
-                    [core :as mg])
+  (:require [medley.core :as m]
             [metabase.db :refer :all]
-            [metabase.driver :as driver]
-            [metabase.driver.mongo.util :refer [with-mongo-connection]]
             (metabase.models [database :refer [Database]]
                              [field :refer [Field]]
                              [table :refer [Table]])
-            [metabase.test-data.data :as data]))
-
-(declare load-data
-         set-field-types!)
-
-;; ## CONSTANTS
-
-(def ^:private ^:const mongo-test-db-conn-str
-  "Connection string for the Metabase Mongo Test DB." ; TODO - does this need to come from an env var so it works in CircleCI?
-  "mongodb://localhost/metabase-test")
-
-(def ^:private ^:const mongo-test-db-name
-  "Name of the Mongo test database."
-  "Mongo Test")
-
+            [metabase.test.data :refer [get-or-create-database!]]
+            (metabase.test.data [data :as data]
+                                [mongo :as loader])))
 
 ;; ## MONGO-TEST-DB + OTHER DELAYS
-
-(defn destroy!
-  "Remove `Database`, `Tables`, and `Fields` for the Mongo test DB."
-  []
-  #_(cascade-delete Database :name mongo-test-db-name))
 
 (defonce
   ^{:doc "A delay that fetches or creates the Mongo test `Database`.
           If DB is created, `load-data` and `sync-database!` are called to get the DB in a state that we can use for testing."}
   mongo-test-db
-  (delay
-   ;; Resolve metabase.test.data.datasets at runtime to avoid circular dependency
-   (require 'metabase.test.data.datasets)
-   (assert (contains? @@(ns-resolve 'metabase.test.data.datasets 'test-dataset-names) :mongo)
-           "Why are we attempting to use the Mongo test Database when we're not testing against mongo?")
-   (let [db (or (sel :one Database :name mongo-test-db-name)
-                (let [db (ins Database
-                           :name mongo-test-db-name
-                           :engine :mongo
-                           :details {:conn_str mongo-test-db-conn-str})]
-                  (log/info (color/cyan "Loading Mongo test data..."))
-                  (load-data)
-                  (driver/sync-database! db)
-                  (set-field-types!)
-                  (log/info (color/cyan "Done."))
-                  db))]
-     (assert (and (map? db)
-                  (integer? (:id db))
-                  (exists? Database :id (:id db))))
-     db)))
+  (delay (get-or-create-database! (loader/dataset-loader) data/test-data)))
 
 (defonce
   ^{:doc "A Delay that returns the ID of `mongo-test-db`, forcing creation of it if needed."}
@@ -68,6 +26,7 @@
 
 
 ;; ## FNS FOR GETTING RELEVANT TABLES / FIELDS
+;; TODO - This seems like it's duplicated a bit with the functions in metabase.test.data
 
 (defn table-name->table
   "Fetch `Table` for Mongo test database.
@@ -115,35 +74,3 @@
             (keyword? field-name)]
       :post [(integer? %)]}
      (:id (field-name->field table-name field-name)))))
-
-
-;; ## LOADING STUFF
-
-(defn- load-data
-  "Load the data for the Mongo test database. This can safely be called multiple times; duplicate documents will *not* be inserted."
-  []
-  (with-mongo-connection [mongo-db @mongo-test-db]
-    (doseq [[collection {fields :fields rows :rows}] data/test-data]
-      (let [fields (map :name fields)
-            rows (map-indexed (partial vector) rows)]
-        (doseq [[i row] rows]
-          (try
-            (mc/insert mongo-db (name collection) (assoc (zipmap fields row)
-                                                         :_id (inc i)))
-            (catch com.mongodb.MongoException$DuplicateKey _)))
-        (log/info (color/cyan (format "Loaded data for collection '%s'." (name collection))))))))
-
-(defn- set-field-types! []
-  (doseq [[collection-name {fields :fields}] data/test-data]
-    (doseq [{:keys [special-type field-type] :as field} fields]
-      (when (or field-type special-type)
-        (let [table-id (sel :one :id Table :name (name collection-name))
-              _        (assert (integer? table-id))
-              field-id (sel :one :id Field :table_id table-id :name (name (:name field)))
-              _        (assert (integer? table-id))]
-          (when special-type
-            (log/info (format "SET SPECIAL TYPE %s.%s -> %s..." collection-name (:name field) special-type))
-            (upd Field field-id :special_type special-type))
-          (when field-type
-            (log/info (format "SET FIELD TYPE %s.%s -> %s..." collection-name (:name field) field-type))
-            (upd Field field-id :field_type field-type)))))))
